deck layer is detached
