--- web-client/src/stores/deckStore.ts
@@ -1,16 +1,0 @@
-import { defineStore } from 'pinia';
-import { Deck } from '@deck.gl/core/typed';
-
-export const useDeckStore = defineStore('deck', {
-    state: () => ({
-        deckInstance: null as any,
-    }),
-    actions: {
-        setDeckInstance(instance:Deck) {
-            this.deckInstance = instance;
-        },
-        getDeckInstance() {
-            return this.deckInstance;
-        }
-    },
-});
