+++ web-client/src/views/AdvancedUserView.vue
@@ -6,7 +6,6 @@
     import { onMounted } from 'vue';
     import { useAdvancedModeStore } from '@/stores/advancedModeStore.js';
     import { useMapStore } from '@/stores/mapStore';
-    import { createDeckGLInstance} from '@/composables/SrMapUtils';
     import { Map as OLMap } from 'ol';
     import SrModeSelect from "@/components/SrModeSelect.vue";
  
@@ -22,20 +21,7 @@
         console.log(`Current root font size: ${fontSize}`);
 
         advancedModeStore.advanced = true;
-        const mapStore = useMapStore();
-        const map = mapStore.getMap() as OLMap ;
-        if (map){
-            const tgt = map.getViewport() as HTMLDivElement; 
-            const deckLayer = createDeckGLInstance(tgt);
-            if(deckLayer){
-                map.addLayer(deckLayer);
-                console.log('deckLayer added:',deckLayer);
-            } else {
-                console.error('createDeckGLInstance returned null');
-            }
-        } else {
-            console.error('map is null');
-        }
+
         console.log('AdvancedUserView onMounted');
     });
 
@@ -54,7 +40,9 @@
             </SrSideBarLayout>
         </template>
         <template v-slot:main>
-            <SrMap />
+            <div class="sr-map-container" ID="AdvUserMapDiv">
+                <SrMap />
+            </div>
         </template>
     </TwoColumnLayout>
 </template>
